chore:Updating method name to OrderGoToTerra and renaming package for internal schema
