+++ internal/schema/fabric_common_schemas.go
@@ -1,4 +1,4 @@
-package equinix_schema
+package schema
 
 import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
